[#158103296]: Start moving things to 'mixins' folder
